requirements.txt
crewai==0.41.1
crewai_tools
firecrawl-py
python-dotenv
langchain_google_genai